Method to create empty node info
